feat: enhance ResourceGroup and UserGroup classes with generics; refactor TripletexAPI to use groups, remove outdated cruds and resource utility functions. Starting example hirarchy for tripletex code
