+++ .gitignore
@@ -301,9 +301,6 @@ Temporary Items
 # ruff
 .ruff_cache/
 
-# LSP config files
-pyrightconfig.json
-
 ### VisualStudioCode ###
 .vscode/*
 !.vscode/settings.json
