+++ pytest.ini
@@ -13,5 +13,6 @@ markers =
 
 filterwarnings =
     ignore:This process.*is multi-threaded.*fork.*may lead to deadlocks:DeprecationWarning
+    ignore:Rate limiting is an EXPERIMENTAL feature.*:FutureWarning
