+++ tests/integration/tripletex_resources/setup.py
@@ -5,8 +5,9 @@ from dotenv import load_dotenv
 
 from crudclient.response_strategies import ModelDumpable
 
+from .api import TripletexAPI
 from .auth import TripletexAuthStrategy
-from .client import TripletexAPI, TripletexClient
+from .client import TripletexClient
 from .config import TripletexConfig, TripletexTestConfig
 from .crud import TripletexCrud
 from .models import (
